Artifact data cleanup job improved the performance of the query (#494)
* Updated cleanup job

* Fixed the mongo query alert for artifact cleanup job

* Added unit test

+++ rest/src/main/java/software/wings/service/intfc/ArtifactService.java
@@ -148,7 +148,7 @@ public interface ArtifactService extends OwnedByApplication {
    *
    * @param retentionSize the size of the artifacts to be retained
    */
-  void deleteArtifacts(long retentionSize);
+  void deleteArtifacts(int retentionSize);
 
   /**
    * Delete by artifact stream.
